Initial release detail work.

+++ RoadieLibrary/Models/Users/UserRatingBase.cs
@@ -7,6 +7,6 @@ namespace Roadie.Library.Models.Users
     {
         public bool IsDisliked { get; set; }
         public bool IsFavorite { get; set; }
-        public int? Rating { get; set; }
+        public short? Rating { get; set; }
     }
 }
